c_n_project/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.354
langchain-core==0.1.3
groq==0.4.1